Move the notifications tests
